+++ src/main/java/net/witcher_rpg/mixin/LivingEntityMixin.java
@@ -3,11 +3,14 @@ package net.witcher_rpg.mixin;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.attribute.DefaultAttributeContainer;
 import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.registry.tag.DamageTypeTags;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
+import net.spell_engine.api.effect.Synchronized;
 import net.spell_engine.internals.SpellRegistry;
 import net.spell_engine.internals.casting.SpellCasterEntity;
+import net.witcher_rpg.WitcherClassMod;
 import net.witcher_rpg.effect.Effects;
 import net.witcher_rpg.entity.attribute.WitcherAttributes;
 import org.spongepowered.asm.mixin.Mixin;
@@ -16,6 +19,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import java.util.Objects;
+import java.util.Random;
 
 import static net.witcher_rpg.WitcherClassMod.MOD_ID;
 
@@ -57,6 +61,26 @@ public abstract class LivingEntityMixin {
         LivingEntity damagedTarget = ((LivingEntity) (Object) this);
         if(damagedTarget.isPlayer() && damagedTarget.hasStatusEffect(Effects.QUEN_ACTIVE) && !source.isIn(DamageTypeTags.BYPASSES_INVULNERABILITY)){
             damagedTarget.heal(amount/2);
+        }
+    }
+    @Inject(method = "damage", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/LivingEntity;applyDamage(Lnet/minecraft/entity/damage/DamageSource;F)V"))
+    private void decreaseAdrenalineAmplifierOnDamage(DamageSource source, float amount, CallbackInfoReturnable<Boolean> cir) {
+        LivingEntity damagedTarget = ((LivingEntity) (Object) this);
+        if(damagedTarget.isPlayer() && damagedTarget.hasStatusEffect(Effects.ADRENALINE_GAIN)){
+            int adrenaline_effect_amplifier = damagedTarget.getStatusEffect(Effects.ADRENALINE_GAIN).getAmplifier();
+            int adrenaline_effect_duration = damagedTarget.getStatusEffect(Effects.ADRENALINE_GAIN).getDuration();
+            float adrenaline_attribute_player = (float) (damagedTarget.getAttributeValue(WitcherAttributes.ADRENALINE_MODIFIER)-100.0F);
+            float random = new Random().nextFloat(100);
+            if(adrenaline_effect_amplifier != 0){
+                if(random > adrenaline_attribute_player){
+                    WitcherClassMod.LOGGER.info("TRUE");
+                    damagedTarget.removeStatusEffect(Effects.ADRENALINE_GAIN);
+                    damagedTarget.addStatusEffect(new StatusEffectInstance(Effects.ADRENALINE_GAIN,adrenaline_effect_duration,adrenaline_effect_amplifier-1,false,false,true));
+                }
+
+            }else{
+                damagedTarget.removeStatusEffect(Effects.ADRENALINE_GAIN);
+            }
         }
     }
 }