+++ src/main/java/net/witcher_rpg/mixin/LivingEntityMixin.java
@@ -7,10 +7,8 @@ import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.registry.tag.DamageTypeTags;
 import net.minecraft.server.network.ServerPlayerEntity;
 import net.minecraft.util.Identifier;
-import net.spell_engine.api.effect.Synchronized;
 import net.spell_engine.internals.SpellRegistry;
 import net.spell_engine.internals.casting.SpellCasterEntity;
-import net.witcher_rpg.WitcherClassMod;
 import net.witcher_rpg.effect.Effects;
 import net.witcher_rpg.entity.attribute.WitcherAttributes;
 import org.spongepowered.asm.mixin.Mixin;
@@ -73,9 +71,9 @@ public abstract class LivingEntityMixin {
             float random = new Random().nextFloat(100);
             if(adrenaline_effect_amplifier != 0){
                 if(random > adrenaline_attribute_player){
-                    WitcherClassMod.LOGGER.info("TRUE");
                     damagedTarget.removeStatusEffect(Effects.ADRENALINE_GAIN);
-                    damagedTarget.addStatusEffect(new StatusEffectInstance(Effects.ADRENALINE_GAIN,adrenaline_effect_duration,adrenaline_effect_amplifier-1,false,false,true));
+                    damagedTarget.addStatusEffect(new StatusEffectInstance(Effects.ADRENALINE_GAIN,adrenaline_effect_duration,
+                            adrenaline_effect_amplifier-1,false,false,true));
                 }
 
             }else{