+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - added more configured_feature variants for the witcher ores
 - tweaked some numbers for the ores
 - made the dark iron ore brighter
+- added a meteorite like geode
 
 1.1.2
 - changed the aard impact particles to wind vacuum for better a better shockwave feeling